% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_merging.R
\name{generic_region_merging}
\alias{generic_region_merging}
\title{Call OTB GenericRegionMerging.}
\usage{
generic_region_merging(
  input,
  ofile = tempfile(fileext = ".tif"),
  thresh = 10,
  spec = 0.1,
  spat = 0.5,
  method = "bs",
  otb_dir = "C:/OTB/bin"
)
}
\arguments{
\item{input}{SpatRaster. Multiband raster to segment}

\item{ofile}{string. The path where to save the output of GRM}

\item{thresh, spec, spat}{numeric. refer to paper or OTB GRM webpage for description of parameters
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{method}{string. refer to paper or OTB GRM webpage for description of parameters
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{otb_dir}{string. Directory location of OTB (where you installed OTB earlier).
Likely "C:/OTB/bin" on Windows}
}
\value{
the raster produced by OTB
}
\description{
Call OTB GenericRegionMerging. OTB must be installed on your computer first.
\url{https://www.orfeo-toolbox.org/CookBook-7.0/index.html}
}
\examples{
\dontrun{
library(terra)
f = system.file("extdata", "metrics.tif", package="pfif")
g = system.file("extdata", "roads.gpkg", package="pfif")
h = system.file("extdata", "stream.gpkg", package="pfif")
metrics = rast(f)
roads = vect(g)
stream = vect(h)

plot(metrics)
plot(metrics[[1]])
plot(roads, add = TRUE)
plot(stream, add = TRUE)

masks = list(roads, stream)

layers = pre_processing(metrics, masks = masks)
plot(layers)

ofile = paste0(tempdir(), "/grm.tif")
grm = generic_region_merging(layers, ofile, otb_dir = "/home/jr/Logiciels/OTB-8.1.2-Linux64/bin")

color = grDevices::colors()[grep('gr(a|e)y', grDevices::colors(), invert = TRUE)]
plot(grm, col= color)

polygons = terra::as.polygons(grm)

plotRGB(layers)
plot(polygons, border = "red", add = TRUE)
}
}
