% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{polygon_statistics}
\alias{polygon_statistics}
\title{Computes summary statistics about the polygons}
\usage{
polygon_statistics(polygons, on = NULL)
}
\arguments{
\item{polygons}{sf object with polygons}

\item{on}{character. A vector of attribute on which to compute the metrics}
}
\description{
Computes summary statistics about the polygons to describe how they are spread, shaped, sized,
homogeneous and so on. By defaut it computes the polygon density (number of polygons per unit area).
Then it compute the Moran's index, the mean and standard deviation of the attributes of interest
recorded with the polygons. It can computes these indices on the 'area' of the polygon and on each
attribute provided by the user.
}
\examples{
library(sf)
f = system.file("extdata", "inventory.gpkg", package="pfif")
inventory = st_read(f)

polygon_statistics(inventory)
polygon_statistics(inventory, on = c("area"))
}
