% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Postprocessing.R
\name{post_processing}
\alias{post_processing}
\title{Post processing}
\usage{
post_processing(grm)
}
\arguments{
\item{grm}{string. path to the raster produced by generic raster merging}
}
\value{
path to vectorized GRM written on disk
}
\description{
Post processing. Vectorize the GRM raster (as far as I can tell)
}
