% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-Preprocessing.R
\name{pre_processing}
\alias{pre_processing}
\title{Pre processing}
\usage{
pre_processing()
}
\value{
The paths of the files written on disk
}
\description{
First step after \code{set_pfif_globals()}. It generates \code{spl_stack.tif} and \code{water_roads_polygons.shp}
in the \verb{temp/} folder of the \code{out_dir} directory.
}
