% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_merging.R
\name{generic_region_merging}
\alias{generic_region_merging}
\alias{kmeans_region_merging}
\alias{kmeans_generic_region_merging}
\title{Segmentation of a raster into polygons}
\usage{
generic_region_merging(
  input,
  ofile = tempfile(fileext = ".tif"),
  thresh = 25,
  spec = 0.1,
  spat = 0.5,
  method = "bs",
  otb_dir = "C:/OTB/bin"
)

kmeans_region_merging(input, polygons, k = 10)

kmeans_generic_region_merging(
  input,
  ofile = tempfile(fileext = ".tif"),
  thresh = 10,
  spec = 0.1,
  spat = 0.5,
  method = "bs",
  k = 10,
  otb_dir = "C:/OTB/bin"
)
}
\arguments{
\item{input}{SpatRaster. Multiband raster to segment}

\item{ofile}{string. The path where to save the output of GRM}

\item{thresh, spec, spat}{numeric. refer to paper or OTB GRM webpage for description of parameters
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{method}{string. refer to paper or OTB GRM webpage for description of parameters
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{otb_dir}{string. Directory location of OTB (where you installed OTB earlier).
Likely "C:/OTB/bin" on Windows}

\item{polygons}{sf object. Vector of polygons}

\item{k}{integer Number of clusters}
}
\value{
the raster produced by OTB
}
\description{
\code{generic_region_merging} is the original and the main function. It calls the OTB GenericRegionMerging software.
OTB must be installed on your computer first. See  \url{https://www.orfeo-toolbox.org/CookBook-7.0/index.html}\cr\cr
\code{kmeans_region_merging} takes the output of \code{generic_region_merging} and aggreates similar polygons into
bigger polygons to generate of more realistic EFI with few and larger polygons using a K-Means Clustering
approach.\cr\cr
\code{kmean_generic_region_merging} is simply both algorithms in a single call.
}
\examples{
\dontrun{
library(terra)
f = system.file("extdata", "metrics.tif", package="pfif")
g = system.file("extdata", "roads.gpkg", package="pfif")
h = system.file("extdata", "stream.gpkg", package="pfif")
metrics = rast(f)
roads = vect(g)
stream = buffer(vect(h), 2)

plot(metrics)
plot(metrics[[1]])
plot(roads, add = TRUE)
plot(stream, add = TRUE)

masks = list(roads, stream)

layers = pre_processing(metrics, masks = masks)
plot(layers)

ofile = paste0(tempdir(), "/grm.tif")
otb_dir = "/home/jr/Logiciels/OTB-8.1.2-Linux64/bin"
polygons = generic_region_merging(layers, ofile, otb_dir = otb_dir)

plotRGB(layers, r = 3, g = 1, b = 4)
plot(polygons, add = TRUE)
}
}
