% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-Configuration.R
\name{set_pfif_globals}
\alias{set_pfif_globals}
\title{Set global parameters}
\usage{
set_pfif_globals(root, out_dir, lc_f, otb_dir = "C:/OTB/bin")
}
\arguments{
\item{root}{string. Root directory with all data from the drive.}

\item{out_dir}{string. Output directory for files generated.}

\item{lc_f}{string. File location of 2018 VLCE 2.0 landcover data using 2018 because it is
the year of Romeo ALS acquisition. Download here: \link{https://opendata.nfis.org/mapserver/nfis-change_eng.html}
\link{https://opendata.nfis.org/downloads/forest_change/CA_forest_VLCE2_2018.zip}}

\item{otb_dir}{string. Directory location of OTB (where you installed OTB earlier). Likely "C:/OTB/bin"
on Windows.}
}
\description{
Set global parameters to make working other functions
}
\details{
When \code{root} points to the directory \verb{RMF_EFI_layers/} it automatically searches for other files
including:
\itemize{
\item{RMF_EFI_layers/Polygons Inventory/RMF_PolygonForest.shp}
\item{RMF_EFI_layers/SPL100 metrics/RMF_20m_T130cm_p95.tif}
\item{RMF_EFI_layers/SPL100 metrics/RMF_20m_T130cm_2m_cov.tif}
\item{RMF_EFI_layers/SPL100 metrics/RMF_20m_T130cm_cv.tif}
\item{RMF_EFI_layers/Roads/RMF_roads.shp}}
}
\examples{
root = "/media/jr/Alexandre/RMF_EFI_layers/"
out_dir <- '/home/jr/Documents/Ulaval/Projets Annexes/Alexandre - ontario'
lc_f <- "/home/jr/Téléchargements/CA_forest_VLCE2_2018/CA_forest_VLCE2_2018.tif"
otb_dir <- "/home/jr/Logiciels/OTB-8.1.2-Linux64/bin"

set_pfif_globals(root, out_dir, lc_f, otb_dir)
}
