% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-GRM.R
\name{generic_region_merging}
\alias{generic_region_merging}
\title{Call OTB GRM}
\usage{
generic_region_merging(thresh = 10, spec = 0.1, spat = 0.5)
}
\arguments{
\item{thresh, spec, spat}{numeric. refer to paper or OTB GRM webpage for description of parameters
\link{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}
}
\value{
the path to the segmented raster file
}
\description{
Call OTB GenericRegionMerging. This takes a lot of time and uses a lot of memory. It may be more
advised to run it in command line outside R
}
