% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{pre_processing}
\alias{pre_processing}
\title{Pre-processing stage}
\usage{
pre_processing(layers, smooth = 5, masks = NULL)
}
\arguments{
\item{layers}{SpatRaster. Multiband raster to segment with \link{generic_region_merging}}

\item{smooth}{odd integer. 0 to skip the smoothing stage}

\item{masks}{List of SpatVector object that will be used to mask the raster}
}
\description{
Using a multiband raster that will later be used as a basis for polygon segmentation, the function masks
the unwanted pixel (rivers, lakes, roads, ...), smooth the layers, and normalize the value to range
in \verb{[0,100]}
}
\examples{
library(terra)

f = system.file("extdata", "metrics.tif", package="pfif")
g = system.file("extdata", "roads.gpkg", package="pfif")
h = system.file("extdata", "stream.gpkg", package="pfif")
metrics = rast(f)
roads = vect(g)
stream = vect(h)

plot(metrics)
plot(metrics[[1]])
plot(roads, add = TRUE)
plot(stream, add = TRUE)

masks = list(roads, stream)

layers = pre_processing(metrics, masks = masks)
plot(layers)
}
